Added net value column to Asset Declarations table

+++ AssetDeclarationsApi/AssetDeclarationsApi/Entities/CashPosition.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Currency { get; set; }
-        public double? CurrencyValue { get; set; }
-        public double? BaseValue { get; set; }
+        public decimal? CurrencyValue { get; set; }
+        public decimal? BaseValue { get; set; }
 
         public int AssetDeclarationId { get; set; }
     }
