+++ AssetDeclarationsApi/AssetDeclarationsApi/Entities/Income.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public double? YearlyValue { get; set; }
+        public decimal? YearlyValue { get; set; }
         public int AssetDeclarationId { get; set; }
     }
 }
