+++ AssetDeclarationsApi/AssetDeclarationsApi/Entities/Liability.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public double? Value { get; set; }
+        public decimal? Value { get; set; }
         public int AssetDeclarationId { get; set; }
     }
 }
