+++ AssetDeclarationsApi/AssetDeclarationsApi/Entities/SecurityPosition.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double? Quantity { get; set; }
-        public double? Value { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? Value { get; set; }
         public int AssetDeclarationId { get; set; }
     }
 }
